refactor(backend): Move files to backend folder
