+++ backend/app/image_renderer/camera_pos_utils.py
